minor code quality fixes

* Removed `systemReset.c`, which appears to have never been implemented.
* Fixed a warning in `arm7/touch.c`, as well as some signedness issues in other files.

+++ source/arm9/console.c
@@ -220,7 +220,7 @@ ssize_t con_write(const char *ptr, size_t len) {
 
 	char chr;
 
-	int i, count = 0;
+	size_t i, count = 0;
 	char *tmp = (char*)ptr;
 	int intensity = 0;
